se agrego nuevas validaciones

+++ src/app/components/color/color.component.ts
@@ -1,21 +1,22 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { CookieService } from 'ngx-cookie-service';
 import { IColor } from 'src/app/interfaces/IColor';
 import { ColorService } from 'src/app/services/color.service';
 import { ToastrService } from 'node_modules/ngx-toastr';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-color',
   templateUrl: './color.component.html',
   styleUrls: ['./color.component.css'],
 })
-export class ColorComponent {
+export class ColorComponent implements OnInit {
   listColor: IColor[] = [];
   loading: boolean = false;
   formColor: FormGroup;
-
+  searchText: any;
 
   constructor(
     private _colorService: ColorService,
@@ -29,28 +30,17 @@ export class ColorComponent {
       name_col: ['', Validators.required],
       state: ['1', Validators.required],
     });
-
-
   }
-
   ngOnInit(): void {
     this.getColors();
-
   }
 
-  handleClear(){
+  handleClear() {
     this.formColor = this.fb.group({
       name_col: ['', Validators.required],
     });
   }
 
-  // redirect() {
-  //   const params = Number(this.aRouter.snapshot.paramMap.get('idcolor'));
-  //   this.id = Number(this.aRouter.snapshot.paramMap.get('idcolor'));
-  //   if (this.id) {
-  //   }
-  //   this.router.navigate(['/color', '{{id}}']);
-  // }
 
   getColors() {
     this.loading = true;
@@ -62,27 +52,29 @@ export class ColorComponent {
     }, 500);
   }
 
-  // getOneColor(id: number) {
-  //   this.loading = true;
-  //   this._colorService.getOneColor(id).subscribe((data: IColor) => {
-  //     this.loading = false;
-  //     this.formColor2.setValue({
-  //       name_col: data.name_col,
-  //       state: data.state,
-  //     });
-  //     console.log(data);
-  //   });
-  // }
-
-
-
   deleteColor(id: number) {
-    this.loading = true;
-    this._colorService.deleteColor(id).subscribe(() => {
-      this.getColors();
-      this.toastr.warning('El color fue eliminado satisfactoriamente');
+    Swal.fire({
+      title: 'Seguro que desea eliminarlo?',
+      text: "Se eliminara el color",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si',
+      cancelButtonText: 'No',
+
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.loading = true;
+        this._colorService.deleteColor(id).subscribe(() => {
+          this.getColors();
+          // this.toastr.warning('El color fue eliminado satisfactoriamente');
+        });
+        this.getColors();
+        Swal.fire('Se Elimino Correctamente', 'success');
+      }
     });
-    this.getColors();
+
 
   }
 
@@ -94,21 +86,6 @@ export class ColorComponent {
     this._colorService.addColor(color).subscribe(() => {
       this.toastr.success('El color se agrego correctamente');
       this.getColors();
-
     });
   }
-
-  // updateColor() {
-  //   const color: IColor = {
-  //     name_col: this.formColor2.value.name_col,
-  //     state: this.formColor2.value.state,
-  //   };
-
-  //   color.idcolor = this.id;
-  //   this._colorService.updateColor(this.id, color).subscribe(() => {
-  //     this.toastr.success('El color se actualizo correctamente');
-  //     this.loading = false;
-  //     this.getColors();
-  //   });
-  // }
 }
